requirements.txt
# discord.py
discord.py-self
python-dotenv
tweepy